% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apde-package.R
\docType{package}
\name{apde-package}
\alias{apde}
\alias{apde-package}
\title{apde: Common functions used in APDE's work}
\description{
Primarily for internal use in Public Health - Seattle & King County's Assessment, Policy Development, and Evaluation (APDE) Unit. Contains common functions used in APDE's work, including for data processing (extract, transform, load), geocoding, and analyses.
}
\keyword{internal}
