% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apde_keyring.R
\name{apde_keyring_set_f}
\alias{apde_keyring_set_f}
\title{Set or Update a Keyring}
\usage{
apde_keyring_set_f(keyring = NA)
}
\arguments{
\item{keyring}{A character string specifying the name of the keyring. If not provided, a dialog will prompt for input.}
}
\value{
A message indicating the keyring and username that have been set.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

apde_keyring_set_f() was deprecated in apde 0.4.4. Please use \code{\link[apde.etl:apde_keyring_set]{apde.etl::apde_keyring_set()}} instead.

This function creates or updates a keyring with a specified username.
}
\examples{
\dontrun{
apde_keyring_set_f("my_keyring")
}

}
