% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notify.R
\name{apde_notify_address_create_f}
\alias{apde_notify_address_create_f}
\title{Create a New APDE Notify Address}
\usage{
apde_notify_address_create_f(address)
}
\arguments{
\item{address}{Character. Email address to create.}
}
\value{
Integer. ID of the newly created address.
}
\description{
Create a New APDE Notify Address
}
\examples{
 \dontrun{
  # ENTER EXAMPLES HERE
 }
}
