% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notify.R
\name{apde_notify_address_delete_f}
\alias{apde_notify_address_delete_f}
\title{Delete an APDE Notify Address}
\usage{
apde_notify_address_delete_f(address_id = NULL, address = NULL)
}
\arguments{
\item{address_id}{Integer. ID of the address to delete.}

\item{address}{Character. Email address to delete.}
}
\value{
None
}
\description{
Delete an APDE Notify Address
}
\examples{
 \dontrun{
  # ENTER EXAMPLES HERE
 }
}
