% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notify.R
\name{apde_notify_address_get_f}
\alias{apde_notify_address_get_f}
\title{Retrieve a Specific APDE Notify Address}
\usage{
apde_notify_address_get_f(address_id = NULL, address = NULL)
}
\arguments{
\item{address_id}{Integer. ID of the address to retrieve.}

\item{address}{Character. Email address to retrieve.}
}
\value{
A dataframe containing the address information.
}
\description{
Retrieve a Specific APDE Notify Address
}
\examples{
 \dontrun{
  # ENTER EXAMPLES HERE
 }
}
