% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notify.R
\name{apde_notify_address_id_get_f}
\alias{apde_notify_address_id_get_f}
\title{Get APDE Notify Address ID}
\usage{
apde_notify_address_id_get_f(address_id = NULL, address = NULL)
}
\arguments{
\item{address_id}{Integer. ID of the address.}

\item{address}{Character. Email address.}
}
\value{
Integer. ID of the address.
}
\description{
Get APDE Notify Address ID
}
\examples{
 \dontrun{
  # ENTER EXAMPLES HERE
 }
}
