% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notify.R
\name{apde_notify_address_set_f}
\alias{apde_notify_address_set_f}
\title{Update an APDE Notify Address}
\usage{
apde_notify_address_set_f(address_id = NULL, address = NULL, new_address)
}
\arguments{
\item{address_id}{Integer. ID of the address to update.}

\item{address}{Character. Current email address.}

\item{new_address}{Character. New email address.}
}
\value{
None
}
\description{
Update an APDE Notify Address
}
\examples{
 \dontrun{
  # ENTER EXAMPLES HERE
 }
}
