% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notify.R
\name{apde_notify_addresses_get_f}
\alias{apde_notify_addresses_get_f}
\title{Retrieve All APDE Notify Addresses}
\usage{
apde_notify_addresses_get_f()
}
\value{
A dataframe containing all notify addresses.
}
\description{
Retrieve All APDE Notify Addresses
}
\examples{
 \dontrun{
  # ENTER EXAMPLES HERE
 }
}
