% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notify.R
\name{apde_notify_f}
\alias{apde_notify_f}
\title{Send APDE Notification}
\usage{
apde_notify_f(msg_id = NULL, msg_name = NULL, vars)
}
\arguments{
\item{msg_id}{Integer. ID of the message to send.}

\item{msg_name}{Character. Name of the message to send.}

\item{vars}{List. Variables to substitute in the message body.}
}
\value{
None
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

apde_notify_f() was deprecated in apde 0.4.4. Please use \code{\link[apde.etl:apde_notify]{apde.etl::apde_notify()}} instead.
}
\examples{
 \dontrun{
  # ENTER EXAMPLES HERE
 }
}
