% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notify.R
\name{apde_notify_list_get_f}
\alias{apde_notify_list_get_f}
\title{Retrieve APDE Notify List for a Message}
\usage{
apde_notify_list_get_f(msg_id = NULL, msg_name = NULL)
}
\arguments{
\item{msg_id}{Integer. ID of the message.}

\item{msg_name}{Character. Name of the message.}
}
\value{
A dataframe containing the notify list for the specified message.
}
\description{
Retrieve APDE Notify List for a Message
}
\examples{
 \dontrun{
  # ENTER EXAMPLES HERE
 }
}
