% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notify.R
\name{apde_notify_list_set_f}
\alias{apde_notify_list_set_f}
\title{Set APDE Notify List for a Message}
\usage{
apde_notify_list_set_f(msg_id = NULL, msg_name = NULL, choices)
}
\arguments{
\item{msg_id}{Integer. ID of the message.}

\item{msg_name}{Character. Name of the message.}

\item{choices}{Vector. List of email addresses to associate with the message.}
}
\value{
None
}
\description{
Set APDE Notify List for a Message
}
\examples{
 \dontrun{
  # ENTER EXAMPLES HERE
 }
}
