% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notify.R
\name{apde_notify_menu_f}
\alias{apde_notify_menu_f}
\title{Launch APDE Notify Menu}
\usage{
apde_notify_menu_f()
}
\value{
A Shiny app object
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

apde_notify_menu_f() was deprecated in apde 0.4.4. Please use \code{\link[apde.etl:apde_notify_menu]{apde.etl::apde_notify_menu()}} instead.

Launches a Shiny app for managing APDE Notify messages and email lists.
}
\examples{
 \dontrun{
  # ENTER EXAMPLES HERE
 }
}
