% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notify.R
\name{apde_notify_msg_get_f}
\alias{apde_notify_msg_get_f}
\title{Retrieve a Specific APDE Notify Message}
\usage{
apde_notify_msg_get_f(msg_id = NULL, msg_name = NULL)
}
\arguments{
\item{msg_id}{Integer. ID of the message to retrieve.}

\item{msg_name}{Character. Name of the message to retrieve.}
}
\value{
A dataframe containing the message information.
}
\description{
Retrieve a Specific APDE Notify Message
}
\examples{
 \dontrun{
  # ENTER EXAMPLES HERE
 }
}
