% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notify.R
\name{apde_notify_msg_id_get_f}
\alias{apde_notify_msg_id_get_f}
\title{Get APDE Notify Message ID}
\usage{
apde_notify_msg_id_get_f(msg_id = NULL, msg_name = NULL)
}
\arguments{
\item{msg_id}{Integer. ID of the message.}

\item{msg_name}{Character. Name of the message.}
}
\value{
Integer. ID of the message.
}
\description{
Get APDE Notify Message ID
}
\examples{
 \dontrun{
  # ENTER EXAMPLES HERE
 }
}
