% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notify.R
\name{apde_notify_msg_set_f}
\alias{apde_notify_msg_set_f}
\title{Set or Update APDE Notify Message}
\usage{
apde_notify_msg_set_f(msg_id = 0, msg_name, msg_subject, msg_body, msg_from)
}
\arguments{
\item{msg_id}{Integer. ID of the message to update. Use 0 for new messages.}

\item{msg_name}{Character. Name of the message.}

\item{msg_subject}{Character. Subject of the message.}

\item{msg_body}{Character. Body of the message.}

\item{msg_from}{Character. Sender's email address.}
}
\value{
Integer. ID of the new or updated message.
}
\description{
Set or Update APDE Notify Message
}
\examples{
 \dontrun{
  # ENTER EXAMPLES HERE
 }
}
