% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notify.R
\name{apde_notify_msgs_get_f}
\alias{apde_notify_msgs_get_f}
\title{Retrieve APDE Notify Messages}
\usage{
apde_notify_msgs_get_f()
}
\value{
A dataframe containing message information
}
\description{
Fetches all parent messages from the APDE notify_msgs table.
}
\examples{
 \dontrun{
  # ENTER EXAMPLES HERE
 }
}
