% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_db_connection.R
\name{create_db_connection}
\alias{create_db_connection}
\title{Create Database Connection Function}
\usage{
create_db_connection(
  server = c("phextractstore", "hhsaw", "inthealth"),
  prod = TRUE,
  interactive = FALSE
)
}
\arguments{
\item{server}{Whether we are connecting to HHSAW, inthealth, or phextractstore}

\item{prod}{Whether the production or dev/WIP servers are to be used. Defaults to \code{prod = TRUE}.}

\item{interactive}{Whether to input password manually for HHSAW/inthealth or to
use keyring. True should be used when connecting on VM, typically, or if
keyring/odbc are not set up yet for a given server. Defaults to \code{interactive = FALSE}.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

create_db_connection() was deprecated in apde 0.4.4. Please use \code{\link[apde.etl:create_db_connection]{apde.etl::create_db_connection()}} instead.

Create a connection to the prod or dev servers that APDE uses
}
\author{
Jeremy Whitehurst

Kai Fukutaki
}
