% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etl_qa_run_pipeline.R
\name{default_value}
\alias{default_value}
\alias{\%||\%}
\title{Helper function to provide a default value for NULL}
\usage{
x \%||\% y
}
\arguments{
\item{x}{The primary value to check}

\item{y}{The default value to use if x is NULL}
}
\value{
The value of x if it's not NULL, otherwise the value of y
}
\description{
#' 
This infix function returns the first argument if it's not NULL,
otherwise it returns the second argument.
}
\examples{
# Basic usage
NULL \%||\% 5  # Returns 5
10 \%||\% 5    # Returns 10

# In a function
f <- function(x = NULL) {
  x \%||\% "default"
}
f()        # Returns "default"
f("value") # Returns "value"

}
\keyword{internal}
