% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etl_qa_run_pipeline.R
\name{etl_qa_final_results}
\alias{etl_qa_final_results}
\title{Final QA results for ETL QA pipeline}
\usage{
etl_qa_final_results(initial_qa_results, config)
}
\arguments{
\item{initial_qa_results}{A list containing the initial QA results.}

\item{config}{An S3 object of class "qa_data_config" containing configuration settings.}
}
\value{
A list containing formatted and combined results that consists of:
\item{missingness}{Structured summary of the proportion of missing data per variable and time point}
\item{values}{Combined table with the frequency of categorical variables and simple statistics for numeric and date / datetime variables}
\item{chi_standards}{Comparison of CHI (Community Health Indicator) variables values with those expected based on \code{rads.data::misc_chi_byvars}}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

etl_qa_final_results() was deprecated in apde 0.4.4. Please use \code{\link[apde.etl:etl_qa_final_results]{apde.etl::etl_qa_final_results()}} instead.

This function processes the initial results from \code{\link{etl_qa_initial_results}}
into a format suitable for reporting and visualization. It is the third step
run by \code{\link{etl_qa_run_pipeline}}.
}
\details{
This is an \emph{internal function} accessible only by use of \code{:::}, for example,
\code{apde:::etl_qa_final_results(...)}.
}
\examples{
\dontrun{

# Step 1: generate a config object 
myconfig <- etl_qa_setup_config(
  data_source_type = 'rads',
  data_params = list(
    function_name = 'get_data_birth',
    time_var = 'chi_year',
    time_range = c(2021, 2022),
    cols = c('chi_age', 'race4', 'birth_weight_grams', 'birthplace_city', 
             'num_prev_cesarean', 'mother_date_of_birth'),
    version = 'final', 
    kingco = FALSE, 
    check_chi = TRUE
  ), 
  output_directory = 'C:/temp/'
)

# Step 2: prep the initial results
initial_results <- etl_qa_initial_results(config = myconfig)

# Step 3: prep the final results
final_results <- etl_qa_final_results(initial_qa_results = initial_results, 
                                      config = myconfig)

# Peek at the tables
head(final_results$missingness)
head(final_results$values)
head(final_results$chi_standards)

}

}
\keyword{internal}
