% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external_table_check.R
\name{external_table_check_f}
\alias{external_table_check_f}
\title{Compare external table to source table to find changes}
\usage{
external_table_check_f(
  conn,
  db = c("inthealth_edw", "inthealth_dwhealth"),
  schema,
  table,
  db_ext = "hhs_analytics_workspace",
  conn_ext,
  schema_ext,
  table_ext,
  sql_display = T,
  sql_file_path = NULL,
  overwrite = T
)
}
\arguments{
\item{conn}{SQL server connection created using \code{odbc} package}

\item{db}{name of database/data warehouse for source table, must be inthealth_edw or inthealth_dwhealth}

\item{schema}{name of schema for source table}

\item{table}{name of source table}

\item{db_ext}{name of database/data warehouse for external table}

\item{conn_ext}{name of the connection to the SQL database with source external}

\item{schema_ext}{name of schema for external table}

\item{table_ext}{name of external table}

\item{sql_display}{show the SQL script in Console}

\item{sql_file_path}{write the SQL script to file}

\item{overwrite}{overwrite sql script, appends if FALSE}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

external_table_check_f() was deprecated in apde 0.4.4. Please use \code{\link[apde.etl:external_table_check]{apde.etl::external_table_check()}} instead.
}
\examples{
 \dontrun{
  # ENTER EXAMPLES HERE
 }
 
}
