% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kc_geocode.R
\name{kc_geocode}
\alias{kc_geocode}
\title{King County geocoder - find address candidates}
\usage{
kc_geocode(
  singleline = NULL,
  street = NULL,
  city = NULL,
  zip = NULL,
  max_return = 10,
  best_result = T
)
}
\arguments{
\item{singleline}{A concatenated address that consists of street, city, and ZIP
in a single field}

\item{street}{The street portion of an address (if specifying address separately)}

\item{city}{The city portion of an address (if specifying address separately)}

\item{zip}{The ZIP portion of an address (if specifying address separately)}

\item{max_return}{The maximum number of matches to return (default is 10)}

\item{best_result}{Whether or not to return the single best result for the 
input. Default hierarchy is PointAddress, StreetAddress, then Postal. Multiple 
options within an address type are decided by score.}
}
\description{
\code{kc_geocode} geocodes a single address using KCGIS's ArcGIS
geocoder web service.
}
\details{
This function takes a single address and runs it through King County 
GIS's ArcGIS four-county geocoder web service, which covers King, Pierce, 
Snohomish, and Kitspap Counties. Locators in this service include street networks,
transit, andparcels, and ZIP codes.
Users can input either a single-line address (preferred as it seems to yield 
fewer false matches) or separate street, city, and ZIP values. 
Returns lon/lat (in EPSG 2926 or NAD_1983_HARN_StatePlane_Washington_North_FIPS_4601_Feet 
projection), score (accuracy of match for that locator), locName (name of the 
locator that returned the result), matchAddr (complete address returned), and 
addressType (match level for the result).
}
\examples{
\dontrun{
kc_geocode(singleline = "401 5TH AVE, SEATTLE, 98105", best_result = F)
kc_geocode(street = "401 5TH AVE", city = "SEATTLE", zip = "98105")
}

}
