% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kc_geocode.R
\name{kc_geocode_batch}
\alias{kc_geocode_batch}
\title{King County geocoder - batch geocoder}
\usage{
kc_geocode_batch(id, singleline = NULL, street = NULL, city = NULL, zip = NULL)
}
\arguments{
\item{id}{A unique ID that can be used to link results to original input 
(required, must be numeric).}

\item{singleline}{A concatenated address that consists of street, city, and ZIP
in a single field}

\item{street}{The street portion of an address (if specifying address separately)}

\item{city}{The city portion of an address (if specifying address separately)}

\item{zip}{The ZIP portion of an address (if specifying address separately)}
}
\description{
\code{kc_geocode_batch} geocodes multiple addresses using KCGIS's ArcGIS
geocoder web service.
}
\details{
This function takes up to 1,000 addresses and runs them through King County 
GIS's ArcGIS four-county geocoder web service, which covers King, Pierce, 
Snohomish, and Kitspap Counties. Locators in this service include street networks,
transit, andparcels, and ZIP codes.
Users can input either a single-line address (preferred as it seems to yield 
fewer false matches) or separate street, city, and ZIP values. 
Returns lon/lat (in EPSG 2926 or NAD_1983_HARN_StatePlane_Washington_North_FIPS_4601_Feet 
projection), score (accuracy of match for that locator), locName (name of the 
locator that returned the result), matchAddr (complete address returned), and 
addressType (match level for the result).
}
\examples{
\dontrun{
kc_geocode_batch(id = 1:3,
singleline = c("401 5TH AVE, SEATTLE, 98105", 
"14350 SE EASTGATE WAY, BELLEVUE, 98007",
"33431 13TH PLACE S, FEDERAL WAY, 98003")
}

}
