% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_duplicate.R
\name{table_duplicate_delete_f}
\alias{table_duplicate_delete_f}
\title{Delete tables based on suffix}
\usage{
table_duplicate_delete_f(conn, delete_table_suffix = "_dupe_table_to_delete")
}
\arguments{
\item{conn}{Name of the connection to the FROM SQL database}

\item{delete_table_suffix}{String of a suffix in tables that will be deleted}
}
\value{
None
}
\description{
Delete tables based on suffix
}
\details{
Plans for future improvement ...
}
\examples{
 \dontrun{
  # ENTER EXAMPLES HERE
 }
 
}
\author{
Jeremy Whitehurst, 2024-05-01
}
