% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apde_keyring.R
\name{apde_keyring_check_f}
\alias{apde_keyring_check_f}
\title{Check and Create Keyring if Needed}
\usage{
apde_keyring_check_f(keyring)
}
\arguments{
\item{keyring}{A character string specifying the name of the keyring.}
}
\value{
A message indicating whether the keyring exists or has been created.
}
\description{
This function checks if a keyring exists and creates it if it does not.
}
\examples{
\dontrun{
apde_keyring_check_f("my_keyring")
}

}
