% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notify.R
\name{apde_notify_f}
\alias{apde_notify_f}
\title{Send APDE Notification}
\usage{
apde_notify_f(msg_id = NULL, msg_name = NULL, vars)
}
\arguments{
\item{msg_id}{Integer. ID of the message to send.}

\item{msg_name}{Character. Name of the message to send.}

\item{vars}{List. Variables to substitute in the message body.}
}
\value{
None
}
\description{
Send APDE Notification
}
\examples{
 \dontrun{
  # ENTER EXAMPLES HERE
 }
}
