% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notify.R
\name{apde_notify_menu_f}
\alias{apde_notify_menu_f}
\title{Launch APDE Notify Menu}
\usage{
apde_notify_menu_f()
}
\value{
A Shiny app object
}
\description{
Launches a Shiny app for managing APDE Notify messages and email lists.
}
\examples{
 \dontrun{
  # ENTER EXAMPLES HERE
 }
}
