% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notify.R
\name{apde_notify_set_cred_f}
\alias{apde_notify_set_cred_f}
\title{Set Outlook Credentials for APDE Notify}
\usage{
apde_notify_set_cred_f()
}
\value{
None
}
\description{
Creates and stores Outlook credentials for use with APDE Notify functions.
}
\examples{
 \dontrun{
  # ENTER EXAMPLES HERE
 }
}
