% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_duplicate.R
\name{table_duplicate_f}
\alias{table_duplicate_f}
\title{copy a (smaller) SQL table from server to another}
\usage{
table_duplicate_f(
  conn_from,
  conn_to,
  server_to,
  db_to,
  table_df = data.frame(),
  from_schema = NULL,
  from_table = NULL,
  to_schema = NULL,
  to_table = NULL,
  to_table_prefix = NULL,
  confirm_tables = T,
  delete_table = F,
  delete_table_suffix = "_dupe_table_to_delete"
)
}
\arguments{
\item{conn_from}{Name of the connection to the FROM SQL database}

\item{conn_to}{Name of the connection to the TO SQL database}

\item{server_to}{Name of the server/odbc for the TO SQL database. 
***IMPORTANT*** For any Azure based duplication, there must be a keyring with 
a service name that matches the server_to parameter}

\item{db_to}{Name of the TO SQL database}

\item{table_df}{data.frame that can hold a list of FROM schema, FROM tables 
(required), TO schema, and TO tables}

\item{from_schema}{Variable for FROM schema for either a single table or can 
populate any missing data in the table_df from_schema column}

\item{from_table}{Variable for FROM table for duplicating a single table}

\item{to_schema}{Variable for TO schema for either a single table or can populate 
any missing data in the table_df to_schema column, can be blank and use the 
from_schema on TO database}

\item{to_table}{Variable for TO table for either a single table or can populate 
any missing data in the table_df to_table column, can be blank and use the 
from_table on TO database}

\item{to_table_prefix}{Variable of a prefix to be added before the TO table name}

\item{confirm_tables}{If \code{TRUE}, will require user confirm the list of 
tables being duplicated}

\item{delete_table}{If \code{TRUE}, will delete old TO tables, if \code{FALSE}, 
will rename old TO tables}

\item{delete_table_suffix}{Variable of a suffix to be added after the name of 
old TO tables}
}
\value{
None
}
\description{
Made for smaller table, e.g., reference tables
}
\details{
Plans for future improvement ...
}
\examples{
 \dontrun{
  # ENTER EXAMPLES HERE
 }
 
}
\author{
Jeremy Whitehurst, 2024-05-01
}
